use interfaces

+++ src/core/ecs/SystemManager.ts
@@ -1,9 +1,8 @@
-import { System } from "./System";
-import { World } from "./World";
+import { ISystem } from "./ISystem";
+import { IWorld } from "./IWorld";
 import { IEventEmitter } from "../events/IEventEmitter";
 import { EventEmitter } from "../events/EventEmitter";
 import { ISystemManager } from "./ISystemManager";
-import { IWorld } from "./IWorld";
 
 /**
  * Manages all systems in the ECS architecture.
@@ -13,7 +12,7 @@ export class SystemManager implements ISystemManager {
   /**
    * The collection of registered systems.
    */
-  private systems: System[] = [];
+  private systems: ISystem[] = [];
 
   /**
    * Event emitter for system-related events.
@@ -27,11 +26,11 @@ export class SystemManager implements ISystemManager {
    * @param system The system to register
    * @param world The world the system is being registered with
    */
-  public registerSystem(system: System, world?: IWorld): void {
+  public registerSystem(system: ISystem, world?: IWorld): void {
     this.systems.push(system);
 
     // Call the system's onRegister method if a world is provided
-    if (world) {
+    if (world && system.onRegister) {
       system.onRegister(world);
     }
 
@@ -43,7 +42,7 @@ export class SystemManager implements ISystemManager {
    *
    * @param callback The callback function
    */
-  public onSystemRegistered(callback: (system: System) => void): void {
+  public onSystemRegistered(callback: (system: ISystem) => void): void {
     this.eventEmitter.on("systemRegistered", callback);
   }
 
@@ -66,10 +65,10 @@ export class SystemManager implements ISystemManager {
    * @param systemType The constructor of the system type
    * @returns The system instance, or undefined if not found
    */
-  public getSystem<T extends System>(
+  public getSystem<T extends ISystem>(
     systemType: new (...args: any[]) => T
   ): T | undefined {
-    return this.systems.find((system) => system instanceof systemType) as T;
+    return this.systems.find((system) => system instanceof (systemType as any)) as T;
   }
 
   /**
@@ -80,11 +79,11 @@ export class SystemManager implements ISystemManager {
    * @param world The world the system is being removed from
    * @returns True if the system was removed, false if it wasn't found
    */
-  public removeSystem(system: System, world?: IWorld): boolean {
+  public removeSystem(system: ISystem, world?: IWorld): boolean {
     const index = this.systems.indexOf(system);
     if (index !== -1) {
       // Call the system's onRemove method if a world is provided
-      if (world) {
+      if (world && system.onRemove) {
         system.onRemove(world);
       }
 
@@ -104,11 +103,22 @@ export class SystemManager implements ISystemManager {
     if (world) {
       // Call onRemove for each system
       this.systems.forEach((system) => {
-        system.onRemove(world);
+        if (system.onRemove) {
+          system.onRemove(world);
+        }
       });
     }
 
     this.systems = [];
+  }
+
+  /**
+   * Gets all registered systems.
+   *
+   * @returns An array of all registered system instances
+   */
+  public getAllSystems(): ISystem[] {
+    return this.systems;
   }
 
   /**
@@ -117,7 +127,7 @@ export class SystemManager implements ISystemManager {
    * @param system The system that was registered
    * @private
    */
-  private emitSystemRegistered(system: System): void {
+  private emitSystemRegistered(system: ISystem): void {
     this.eventEmitter.emit("systemRegistered", system);
   }
 }
